requirements.txt
# fastapi
flask[async]
kafka-python
open-meteo
# requests
# grequests
httpx
